requirements_dev.txt
Django==1.7
Jinja2==2.7.3
MarkupSafe==0.23
MySQL-python==1.2.5
PyYAML==3.10
Pygments==1.6
SQLAlchemy==0.9.7
Sphinx==1.2.2
argh==0.23.3
argparse==1.2.1
coverage==3.7.1
csvkit==0.8.0
dbf==0.94.003
docutils==0.12
hurry.filesize==0.9
jdcal==1.0
livereload==2.0.0
openpyxl==2.0.4
pathtools==0.1.2
pep8==1.5.7
psycopg2==2.5.4
py==1.4.22
pyflakes==0.8.1
python-coveralls==2.4.2
python-dateutil==1.5
requests==2.3.0
sh==1.09
six==1.7.3
sphinx-autobuild==0.2.3
tornado==3.1.1
tox==1.7.2
virtualenv==1.11.6
watchdog==0.6.0
xlrd==0.9.3
clint==0.4.1
